+++ hondana/author.py
@@ -138,7 +138,7 @@ class Author:
         self.version: int = self._attributes["version"]
         self._created_at = self._attributes["createdAt"]
         self._updated_at = self._attributes["updatedAt"]
-        self._manga_relationships: list[MangaResponse] = relationship_finder(relationships, "manga")  # type: ignore - can't narrow this
+        self._manga_relationships: list[MangaResponse] = relationship_finder(relationships, "manga")  # type: ignore # can't narrow this
         self.__manga: Optional[list[Manga]] = None
 
     def __repr__(self) -> str:
@@ -242,7 +242,7 @@ class Author:
         from .manga import Manga
 
         for manga_id in ids:
-            data = await self._http._view_manga(manga_id, includes=MangaIncludes())
+            data = await self._http._get_manga(manga_id, includes=MangaIncludes())
             formatted.append(Manga(self._http, data["data"]))
 
         if not formatted: