+++ hondana/utils.py
@@ -29,6 +29,7 @@ import json
 import logging
 import pathlib
 import re
+import warnings
 from enum import Enum
 from functools import wraps
 from typing import (
@@ -49,6 +50,14 @@ from typing import (
 from urllib.parse import quote as _uriquote
 
 import aiohttp
+
+
+try:
+    import orjson
+except ModuleNotFoundError:
+    HAS_ORJSON = False
+else:
+    HAS_ORJSON = True
 
 from .errors import AuthenticationRequired
 
@@ -77,6 +86,7 @@ __all__ = (
     "to_json",
     "json_or_text",
     "php_query_builder",
+    "deprecated",
     "to_snake_case",
     "to_camel_case",
     "get_image_mime_type",
@@ -233,6 +243,36 @@ def require_authentication(func: Callable[Concatenate[C, B], T]) -> Callable[Con
     return wrapper
 
 
+def deprecated(
+    alternate: Optional[str] = None,
+) -> Callable[[Callable[B, T]], Callable[B, T]]:
+    """A decorator to mark a method as deprecated.
+
+    Parameters
+    -----------
+    alternate: Optional[:class:`str`]
+        The alternate method to use.
+    """
+
+    def decorator(func: Callable[B, T]) -> Callable[B, T]:
+        @wraps(func)
+        def wrapper(*args: B.args, **kwargs: B.kwargs) -> T:
+            warnings.simplefilter("always", DeprecationWarning)  # turn off filter.
+            if alternate is not None:
+                fmt = "{0.__qualname__} is deprecated, use {1} instead."
+            else:
+                fmt = "{0.__qualname__} is deprecated."
+
+            warnings.warn(fmt.format(func, alternate), stacklevel=2, category=DeprecationWarning)
+
+            warnings.simplefilter("default", DeprecationWarning)  # reset filter.
+            return func(*args, **kwargs)
+
+        return wrapper
+
+    return decorator
+
+
 def calculate_limits(limit: int, offset: int, *, max_limit: int = 100) -> tuple[int, int]:
     """A helper function that will calculate the offset and limit parameters for API endpoints.
 
@@ -266,9 +306,21 @@ def calculate_limits(limit: int, offset: int, *, max_limit: int = 100) -> tuple[
     return new_limit, new_offset
 
 
-def to_json(obj: Any) -> str:
-    """A quick object that dumps a Python type to JSON object."""
-    return json.dumps(obj, separators=(",", ":"), ensure_ascii=True)
+if HAS_ORJSON is True:
+
+    def to_json(obj: Any) -> str:
+        """A quick method that dumps a Python type to JSON object."""
+        return orjson.dumps(obj).decode("utf-8")
+
+    _from_json = orjson.loads  # type: ignore # this is guarded in an if.
+
+else:
+
+    def to_json(obj: Any) -> str:
+        """A quick method that dumps a Python type to JSON object."""
+        return json.dumps(obj, separators=(",", ":"), ensure_ascii=True)
+
+    _from_json = json.loads
 
 
 async def json_or_text(response: aiohttp.ClientResponse) -> Union[dict[str, Any], str]:
@@ -277,7 +329,7 @@ async def json_or_text(response: aiohttp.ClientResponse) -> Union[dict[str, Any]
     try:
         if response.headers["content-type"] == "application/json":
             try:
-                return json.loads(text)
+                return _from_json(text)
             except json.JSONDecodeError:
                 pass
     except KeyError:
@@ -401,27 +453,30 @@ def relationship_finder(relationships: list[RT], relationship_type: RelType, *, 
 
 
 @overload
-def relationship_finder(relationships: list[RT], relationship_type: RelType, *, limit: Optional[int]) -> list[RT]:
+def relationship_finder(relationships: list[RT], relationship_type: RelType, *, limit: Optional[int] = ...) -> list[RT]:
     ...
 
 
 def relationship_finder(
     relationships: list[RT], relationship_type: RelType, *, limit: Optional[int] = None
 ) -> Optional[Union[list[RT], RT]]:
-    ret: list[RT] = []
+    if not relationships:
+        return
 
-    for relationship in relationships:
+    ret: list[RT] = []
+    relationships_copy = relationships.copy()
+
+    for relationship in relationships_copy:
         if relationship["type"] == relationship_type:
             if limit == 1:
+                relationships.remove(relationship)
                 return relationship
             else:
                 ret.append(relationship)
+            relationships.remove(relationship)
 
     if limit is not None:
         return ret[:limit]
-
-    if limit == 1:
-        return
 
     return ret